Change class Amalgum to struct Amalgum in a couple files Updated ChiliWin.h with the version from his 3D_Fundamentals framework Added preprocessor directive to avoid redefinition of the WinAPI macros Cast floats to ints for draw calls to get rid of conversion warnings Change min/max to std::min/std::max after updating ChiliWin.h, adding preprocessor directives and adding #include ChiliWin.h to Graphics.h

+++ Engine/Graphics.cpp
@@ -367,7 +367,7 @@ void Graphics::DrawCircle( int Cx, int Cy, int Radius, Color C )
 
 	for( int x = 0; x <= x_pivot; x++ )
 	{
-		int y = ( int )( sqrtf( ( radius_squared - ( x*x ) ) ) + 0.5f );
+		int y = static_cast<int>( sqrtf( static_cast<float>( radius_squared - ( x*x ) ) ) + 0.5f );
 		DrawQuadrant( x, y );
 		DrawQuadrant( y, x );
 	}
